+++ core/src/plugin/tools.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -26,7 +26,7 @@ const toolBuildSchema = createSchema({
   keys: () => ({
     platform: joi
       .string()
-      .allow("darwin", "linux", "windows")
+      .allow("darwin", "linux", "alpine", "windows")
       .required()
       .example("linux")
       .description("The platform this build is for."),