+++ core/src/plugins/container/build.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -155,14 +155,21 @@ export function getDockerBuildArgs(version: string, specBuildArgs: PrimitiveMap)
     ...specBuildArgs,
   }
 
-  return Object.entries(buildArgs).map(([key, value]) => {
-    // 0 is falsy
-    if (value || value === 0) {
-      return `${key}=${value}`
-    } else {
-      // If the value of a build-arg is null, Docker pulls it from
-      // the environment: https://docs.docker.com/engine/reference/commandline/build/
-      return key
-    }
-  })
+  return Object.entries(buildArgs)
+    .map(([key, value]) => {
+      // If the value is empty, we simply don't pass it to docker
+      if (value === "") {
+        return undefined
+      }
+
+      // 0 is falsy
+      if (value || value === 0) {
+        return `${key}=${value}`
+      } else {
+        // If the value of a build-arg is null, Docker pulls it from
+        // the environment: https://docs.docker.com/engine/reference/commandline/build/
+        return key
+      }
+    })
+    .filter((x): x is string => !!x)
 }