+++ core/src/plugins/kubernetes/helm/action.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -20,13 +20,13 @@ import { getHelmDeployStatus } from "./status.js"
 import { posix } from "path"
 import { k8sContainerStopSync } from "../container/sync.js"
 import { helmGetSyncStatus, helmStartSync } from "./sync.js"
-import { makeDocsLink } from "../../../docs/common.js"
+import { makeDocsLinkPlain } from "../../../docs/common.js"
 import { helmVersion } from "./helm-cli.js"
 
 export const getHelmDeployDocs = () => dedent`
   Specify a Helm chart (either in your repository or remote from a registry) to deploy.
 
-  Refer to the [Helm guide](${makeDocsLink`k8s-plugins/actions/deploy/helm`}) for usage instructions.
+  Refer to the [Helm guide](${makeDocsLinkPlain`k8s-plugins/actions/deploy/helm`}) for usage instructions.
 
   Garden uses Helm ${helmVersion}.
 `