+++ core/src/plugins/kubernetes/jib-container.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -168,7 +168,7 @@ async function buildAndPushViaRemote(params: BuildActionParams<"build", Containe
 
     const { log: skopeoLog } = await runner.exec({
       log,
-      command: ["sh", "-c", syncCommand.join(" ")],
+      command: syncCommand,
       timeoutSec: pushTimeout + 5,
       containerName: utilContainerName,
       buffer: true,