+++ core/src/plugins/kubernetes/run.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -44,6 +44,7 @@ import type { RunResult } from "../../plugin/base.js"
 import { LogLevel } from "../../logger/logger.js"
 import { getResourceEvents } from "./status/events.js"
 import stringify from "json-stringify-safe"
+import { commandListToShellScript } from "../../util/escape.js"
 
 // ref: https://kubernetes.io/docs/reference/labels-annotations-taints/#kubectl-kubernetes-io-default-container
 export const K8_POD_DEFAULT_CONTAINER_ANNOTATION_KEY = "kubectl.kubernetes.io/default-container"
@@ -452,7 +453,7 @@ exec 2<&-
 exec 1<>/tmp/output
 exec 2>&1
 
-${cmd.join(" ")}
+${commandListToShellScript(cmd)}
 `
 }
 
@@ -466,16 +467,31 @@ ${cmd.join(" ")}
  */
 function getArtifactsTarScript(artifacts: ArtifactSpec[]) {
   const directoriesToCreate = artifacts.map((a) => a.target).filter((target) => !!target && target !== ".")
-  const tmpPath = "/tmp/.garden-artifacts-" + randomString(8)
+  const tmpPath = commandListToShellScript(["/tmp/.garden-artifacts-" + randomString(8)])
 
-  // TODO: escape the user paths somehow?
+  const createDirectoriesCommands = directoriesToCreate.map((target) =>
+    commandListToShellScript(["mkdir", "-p", target])
+  )
+
+  const copyArtifactsCommands = artifacts.map(({ source, target }) => {
+    const escapedTarget = commandListToShellScript([target || "."])
+
+    // Allow globs (*) in the source path
+    // Note: This works because `commandListToShellScript` wraps every parameter in single quotes, escaping contained single quotes.
+    // The string `bin/*` will be transformed to `'bin/*'` by `commandListToShellScript`. The shell would treat `*` as literal and not expand it.
+    // `replaceAll` transforms that string then to `'bin/'*''`, which allows the shell to expand the glob, everything else is treated as literal.
+    const escapedSource = commandListToShellScript([source]).replaceAll("*", "'*'")
+
+    return `cp -r ${escapedSource} ${escapedTarget} >/dev/null || true`
+  })
+
   return `
 rm -rf ${tmpPath} >/dev/null || true
 mkdir -p ${tmpPath}
 cd ${tmpPath}
 touch .garden-placeholder
-${directoriesToCreate.map((target) => `mkdir -p ${target}`).join("\n")}
-${artifacts.map(({ source, target }) => `cp -r ${source} ${target || "."} >/dev/null || true`).join("\n")}
+${createDirectoriesCommands.join("\n")}
+${copyArtifactsCommands.join("\n")}
 tar -c -z -f - . | cat
 rm -rf ${tmpPath} >/dev/null || true
 `
@@ -570,7 +586,7 @@ async function runWithArtifacts({
       // using that (tar is not statically compiled so we can't copy that directly). Keeping this snippet around
       // for that:
       // await runner.exec({
-      //   command: ["sh", "-c", `sed -n 'w ${arcPath}'; chmod +x ${arcPath}`],
+      //   command: ["/bin/sh", "-c", `mkdir -p /.garden/bin/ && sed -n 'w /.garden/bin/arc' && chmod +x /.garden/bin/arc`],
       //   container: containerName,
       //   ignoreError: false,
       //   input: <binary>,
@@ -586,16 +602,11 @@ async function runWithArtifacts({
       })
     }
 
-    // Escape the command, so that we can safely pass it as a single string
-    const cmd = [...command!, ...(args || [])].map((s) => JSON.stringify(s))
-
     try {
-      const commandScript = getCommandExecutionScript(cmd)
-
       const res = await runner.exec({
         // Pipe the output from the command to the /tmp/output pipe, including stderr. Some shell voodoo happening
         // here, but this was the only working approach I could find after a lot of trial and error.
-        command: ["sh", "-c", commandScript],
+        command: ["sh", "-c", getCommandExecutionScript([...command!, ...(args || [])])],
         containerName: mainContainerName,
         log,
         stdout,
@@ -616,8 +627,6 @@ async function runWithArtifacts({
         startedAt,
       })
     }
-
-    const tarScript = getArtifactsTarScript(artifacts)
 
     // Copy the artifacts
     const tmpDir = await tmp.dir({ unsafeCleanup: true })
@@ -646,7 +655,7 @@ async function runWithArtifacts({
         // Tarball the requested files and stream to the above extractor.
         runner
           .exec({
-            command: ["sh", "-c", tarScript],
+            command: ["sh", "-c", getArtifactsTarScript(artifacts)],
             containerName: mainContainerName,
             log,
             stdout: extractor,