+++ plugins/terraform/test/validation.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Garden Technologies, Inc. <info@garden.io>
+ * Copyright (C) 2018-2024 Garden Technologies, Inc. <info@garden.io>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -30,7 +30,7 @@ describe("terraform validation", () => {
         garden,
         log: garden.log,
         args: {},
-        opts: withDefaultGlobalOpts({}),
+        opts: withDefaultGlobalOpts({ resolve: undefined }),
       })
     })
   }